fix OpenGLES 2.0 usage and improve code
This change the GLES shader code to match Version 2.0, further is
the code cleaned up to remove this double code to become right screen
size and now done by matrix on shader.

Also use it the for Kodi M. coming GL headers who currently included on addon
and becomes removed in future after Kodi have it itself.
