+++ depends/osx/glm/glm.txt
@@ -1,0 +1,1 @@
+glm https://github.com/g-truc/glm/archive/47031aa4b7b079716e4cac496819e7f464b6a713.tar.gz
