+++ screensaver.asteroids/resources/shaders/GLES/vert.glsl
@@ -1,12 +1,14 @@
-#version 300 es
+#version 100
 
-uniform mat4 modelViewProjectionMatrix;
-layout (location = 0) in vec4 m_attrpos;
-layout (location = 1) in vec4 m_attrcol;
-out vec4 m_col;
+uniform mat4 u_modelViewProjectionMatrix;
+
+attribute vec4 a_position;
+attribute vec4 a_color;
+
+varying vec4 v_color;
 
 void main()
 {
-   gl_Position = vec4(m_attrpos.x, m_attrpos.y, m_attrpos.z, 1.0);
-   m_col = m_attrcol;
+  gl_Position = u_modelViewProjectionMatrix * a_position;
+  v_color = a_color;
 }
