+++ src/types.h
@@ -53,7 +53,9 @@ const f32 FLOATEPSILON    = 0.00001f;
 #define SQR(_x)          ((_x)*(_x))
 #define DEGTORAD(d)      ((d)*(PI / 180.0f))
 #define RADTODEG(r)      ((r)*(180.0f / PI))
+#ifndef assert
 #define assert(_x)
+#endif
 
 /***************************** C L A S S E S *******************************/
